refactor: extract Input validation to separate file
